Add new test and fix bug
The bug occurs in the add function, and it's effect is the order of the
elements may be wrong. This will cause other operations to fail because
the operations rely on the sorted order of elements.

The bug happens when a element that is smaller than the value being
inserted right after we get the predecessor node. This will cause our
value to be inserted in between predecessor and the element that is
smaller than our value, resulting in an incorrect order.

The fix is we have to do a check to ensure that the node that our
predecessor node points to is smaller than us, and in our CAS we
assert our expectation that the node has not changed.